se adapto el proyecto para que no de errores
