--- EjMetrosCuadrados/src/main/java/com/metrosCuadrados/MetroscuadradosApplication 2.java
@@ -1,13 +1,0 @@
-package com.metrosCuadrados;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MetroscuadradosApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MetroscuadradosApplication.class, args);
-	}
-
-}
