+++ EjMetrosCuadrados/src/main/java/com/metrosCuadrados/RestControllerMetrosCuadrados.java
@@ -1,6 +1,5 @@
 package com.metrosCuadrados;
 
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
